requirements.txt
requests~=2.31.0
python-decouple